fix: fixed vram overflow in clearScreen
ClearScreen was clearing double the space it should have been clearing, because it was using uint16 as a size instead of uint8. This was causing a overflow in sub screen textcontroller and removing the other background layers aswell. This should also improve performance of the screen clearing function.

+++ source/controllers/text/TextController.cpp
@@ -274,7 +274,7 @@ void TextController::drawText(const std::string& text, int startX, int startY, i
 
 void TextController::clearScreen()
 {
-    dmaFillHalfWords(0, videoBuffer, 256 * 256 * sizeof(uint16_t)); // Clear the video buffer
+    dmaFillHalfWords(0, videoBuffer, 256 * 256 * sizeof(uint8_t)); // Clear the video buffer
 }
 
 // Helper Functions ======================================================
